Feature/newchatpro (#8)
* Modify the picture viewer

* Integrate contacts to select components

* Modify the project package structure

* login mvp

* 1

* contact structure

* 1

* Modify the memory leak

* Set up the purse module structure

* Currency to generate random number

* keystore

* update database framework,greendao 3.2

* modify keystore file

* upgrade database to 3.2

* 1

* Change the database

* wallet account manager and currency manager

* finish create wallet

* mergin feature

* update new lib.so

* finish create currency

* new test file

* remote CurrencyBean

* add currency address proto

* finish wallet create

* add wallet proto file

* set Default currency

* add transfer proto

* Complete normal transfer

* repair proto file

* method test

* add proto

* 1

* wallet struct

* 2

* business struct

* Modify the wallet structure

* reparir transfer

* add http request

* add trnsfer address

* pull code

* repair transfer connect user

* fix payment bug

* repair transfer tips

* add proto

* test prikey encrypt

* 1

* 2

* 3

* fix bug

* repair pay view

* fix bug

* 3

* fix bug

* repair address request currency

* 1

* mergin code

* 2

* transfer detail

* 3

* patment

* tranaction detail

* 4

* transaction fee

* repair transfer

* group transaction

* message list show time

* 4

* 6

* create new group

* gather detail

* 5

* Translation purse module annotation

* Modified to create wallet guide

* fix buglu bug

* translate language ch to en

* Modify or method

* database test code

* Modify the old users to create the wallet process

* Offline messages that do not create groups cannot be displayed

* Modify the logic judgment if they create purse

* HomeAction extends Base

* 1

* Modify the old user interface Bug

* Modify some bugs

* repair some bug

* repair bugs

* half random message

* muti transfer

* new message pb upload

* friend caht

* To optimize the scan

* db update msg

* message translate

* message update

* delete not used msg bean

* data migrate

* The login module of code optimization

* 1

* new connect message

+++ app/src/main/java/connect/database/green/dao/ConversionSettingEntityDao.java
@@ -26,7 +26,7 @@ public class ConversionSettingEntityDao extends AbstractDao<ConversionSettingEnt
     public static class Properties {
         public final static Property _id = new Property(0, Long.class, "_id", true, "_id");
         public final static Property Identifier = new Property(1, String.class, "identifier", false, "IDENTIFIER");
-        public final static Property Snap_time = new Property(2, Long.class, "snap_time", false, "SNAP_TIME");
+        public final static Property Snap_time = new Property(2, Integer.class, "snap_time", false, "SNAP_TIME");
         public final static Property Disturb = new Property(3, Integer.class, "disturb", false, "DISTURB");
     }
 
@@ -69,7 +69,7 @@ public class ConversionSettingEntityDao extends AbstractDao<ConversionSettingEnt
             stmt.bindString(2, identifier);
         }
  
-        Long snap_time = entity.getSnap_time();
+        Integer snap_time = entity.getSnap_time();
         if (snap_time != null) {
             stmt.bindLong(3, snap_time);
         }
@@ -94,7 +94,7 @@ public class ConversionSettingEntityDao extends AbstractDao<ConversionSettingEnt
             stmt.bindString(2, identifier);
         }
  
-        Long snap_time = entity.getSnap_time();
+        Integer snap_time = entity.getSnap_time();
         if (snap_time != null) {
             stmt.bindLong(3, snap_time);
         }
@@ -115,7 +115,7 @@ public class ConversionSettingEntityDao extends AbstractDao<ConversionSettingEnt
         ConversionSettingEntity entity = new ConversionSettingEntity( //
             cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0), // _id
             cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1), // identifier
-            cursor.isNull(offset + 2) ? null : cursor.getLong(offset + 2), // snap_time
+            cursor.isNull(offset + 2) ? null : cursor.getInt(offset + 2), // snap_time
             cursor.isNull(offset + 3) ? null : cursor.getInt(offset + 3) // disturb
         );
         return entity;
@@ -125,7 +125,7 @@ public class ConversionSettingEntityDao extends AbstractDao<ConversionSettingEnt
     public void readEntity(Cursor cursor, ConversionSettingEntity entity, int offset) {
         entity.set_id(cursor.isNull(offset + 0) ? null : cursor.getLong(offset + 0));
         entity.setIdentifier(cursor.isNull(offset + 1) ? null : cursor.getString(offset + 1));
-        entity.setSnap_time(cursor.isNull(offset + 2) ? null : cursor.getLong(offset + 2));
+        entity.setSnap_time(cursor.isNull(offset + 2) ? null : cursor.getInt(offset + 2));
         entity.setDisturb(cursor.isNull(offset + 3) ? null : cursor.getInt(offset + 3));
      }
